+++ packages/react/src/constants/device-sizes.constant.ts
@@ -7,11 +7,11 @@ export const sizeNames: Size[] = ['xsmall', 'small', 'medium', 'large'];
 const sizes = {
   xsmall: {
     min: 0,
-    default: 0,
-    max: 0,
+    default: 160,
+    max: 320,
   },
   small: {
-    min: 320,
+    min: 321,
     default: 321,
     max: 640,
   },
@@ -41,21 +41,37 @@ const sizes = {
 export type Sizes = typeof sizes;
 
 export interface Breakpoints {
+  xsmall?: number;
   small?: number;
   medium?: number;
 }
 
-export const getSizesForBreakpoints = ({ small, medium }: Breakpoints) => {
+export const getSizesForBreakpoints = (breakpoints: Breakpoints) => {
   const newSizes = {
     ...sizes, // Note: this helps get the function from sizes
     ...fastClone(sizes), // Note: this helps to get a deep clone of fields like small, medium etc
   };
 
+  if (!breakpoints) {
+    return newSizes;
+  }
+
+  const { xsmall, small, medium } = breakpoints;
+
+  if (xsmall) {
+    const xsmallMin = Math.floor(xsmall / 2);
+    newSizes.xsmall = {
+      max: xsmall,
+      min: xsmallMin,
+      default: xsmallMin + 1,
+    };
+  }
+
   if (!small || !medium) {
     return newSizes;
   }
 
-  const smallMin = Math.floor(small / 2);
+  const smallMin = xsmall ? newSizes.xsmall.max + 1 : Math.floor(small / 2);
   newSizes.small = {
     max: small,
     min: smallMin,